requirements.txt
urllib3==1.26.15
llama-cpp-python
requests
python-dotenv
flask
